+++ .github/workflows/build.yml
@@ -4,7 +4,7 @@ on:
   release:
     types: [published]
     paths:
-      - buildspecs/img/**
+      - docker/**
       - .github/workflows/build.yml
   push:
     branches:
@@ -13,7 +13,7 @@ on:
       - ci
       - tests
     paths:
-      - buildspecs/img/**
+      - docker/**
       - .github/workflows/build.yml
 
 jobs:
@@ -35,7 +35,7 @@ jobs:
         uses: docker/build-push-action@v3
         if: ${{ github.event_name == 'release' }}
         with:
-          context: buildspecs/img
+          context: docker
           push: true
           tags: |
             ghcr.io/${{ github.repository }}:${{ github.event.release.tag_name }}
@@ -44,7 +44,7 @@ jobs:
         uses: docker/build-push-action@v3
         if: ${{ github.event_name == 'push' }}
         with:
-          context: buildspecs/img
+          context: docker
           push: true
           tags: |
             ghcr.io/${{ github.repository }}:latest
