+++ .github/workflows/tests.yml
@@ -11,7 +11,7 @@ on:
     - .gitignore
     - LICENSE.md
     - README.md
-    - buildspecs/img/**
+    - docker/**
     - .github/workflows/build.yml
 
   push:
@@ -24,7 +24,7 @@ on:
       - .gitignore
       - LICENSE.md
       - README.md
-      - buildspecs/img/**
+      - docker/**
       - .github/workflows/build.yml
 
 env:
